Move OccOsiSymbology to own class under utils. Fix calculation of shares for Options contracts.
